Create Smart-Will UI HTML
The Smart Will contract is designed to create time-locked wallets that hold funds until a specified release time or specific conditions are met. The contract provides an easy-to-use user interface for interacting with the contract functions.

Contract Overview:

The contract allows users to create time-locked wallets, release funds from the wallet, set a new release time, and add beneficiaries to the wallet.

Functions and Features:

Create Time Locked Wallet: Users can create a time-locked wallet by specifying the beneficiaries, corresponding amounts, and the release time. The wallet is created by calling the createWallet() function and providing the required inputs.

Release Funds: Once the release time has passed, users can release the funds to the designated beneficiaries. The releaseFunds() function is called to initiate the release process.

Set Release Time: The contract owner or authorized editors can set a new release time if needed. By providing the desired release time, the setReleaseTime() function can modify the release time.

Add Beneficiary: The contract owner or authorized editors can add additional beneficiaries to the wallet. The addBeneficiary() function allows the inclusion of new beneficiaries and their corresponding amounts.

User Interface (UI):

The provided user interface is a web-based interface that enables users to interact with the contract functions easily. It includes the following sections:

Create Time Locked Wallet: This section allows users to specify the beneficiaries, amounts, and release time for creating a new time-locked wallet.

Release Funds: Users can initiate the release of funds from the wallet by clicking the "Release Funds" button.

Set Release Time: The contract owner or authorized editors can set a new release time by providing the desired release time in the input field and clicking the "Set Release Time" button.

Add Beneficiary: The contract owner or authorized editors can add beneficiaries to the wallet by providing the beneficiary's address and the corresponding amount in the input fields and clicking the "Add Beneficiary" button.

It's important to note that the provided UI assumes the presence of a compatible Ethereum provider (such as MetaMask) and the availability of the web3 object for interaction with the blockchain. The contract address and ABI (Application Binary Interface) need to be provided in the UI code for proper interaction with the deployed contract.

Users should review the code, adapt it to their specific needs, and ensure proper security measures are in place when using or deploying the contract and its associated user interface.